requirements.txt
PyGithub==1.58.0
ruamel.yaml==0.17.21